+++ src/Game/src/main.cpp
@@ -1,12 +1,10 @@
 #include "GameLoop.h"
-#include "Wallet.h"
-#include "Store.h"
 #include "Inventory.h"
+#include "ItemStore.h"
 #include "MainView.h"
+#include "Wallet.h"
 #include "notifyMessage.h"
 #include <memory>
-
-
 
 int main()
 {
@@ -14,7 +12,7 @@ int main()
 
     std::unique_ptr<Inventory> inventory = std::make_unique<Inventory>();
     std::unique_ptr<Wallet> wallet = std::make_unique<Wallet>();
-    std::unique_ptr<Store> store = std::make_unique<Store>();
+    std::unique_ptr<ItemStore> store = std::make_unique<ItemStore>();
     std::unique_ptr<notifyMessage> msg = std::make_unique<notifyMessage>();
     std::unique_ptr<MainView> gamescreen = std::make_unique<MainView>(msg.get(), wallet.get(),
                                                                       inventory.get(), store.get());