+++ src/Screen/src/MainView.h
@@ -4,10 +4,11 @@
 #include "AchievementListView.h"
 #include "InputModeView.h"
 #include "InventoryView.h"
+#include "ItemStoreView.h"
 #include "MessageView.h"
 #include "OptionsView.h"
 #include "StatusView.h"
-#include "StoreView.h"
+#include "UpgradeView.h"
 #include "View.h"
 #include <string>
 
@@ -23,18 +24,20 @@ private:
     Wallet *wallet = nullptr;
     notifyMessage *msg = nullptr;
     Inventory *inventory = nullptr;
-    Store *store = nullptr;
+    ItemStore *itemstore = nullptr;
     void switchActiveView(const std::string &input);
 
     /* lower part of screen, can switch between these views */
-    StoreView storeview = StoreView(msg, wallet, inventory, store);
+    ItemStoreView itemstoreview = ItemStoreView(msg, wallet, inventory, itemstore);
     InventoryView inventoryview = InventoryView(inventory);
-    OptionsView optionsview = OptionsView(msg, wallet, inventory, store, saveFile);
+    UpgradeView upgradeview = UpgradeView(inventory);
+    OptionsView optionsview = OptionsView(msg, wallet, inventory, itemstore, saveFile);
     AchievementListView achievementlistview = AchievementListView(wallet, msg);
-    View *activeView = dynamic_cast<View *>(&storeview);
+    View *activeView = dynamic_cast<View *>(&itemstoreview);
     std::vector<View *> allViews = {
-        dynamic_cast<View *>(&storeview),
+        dynamic_cast<View *>(&itemstoreview),
         dynamic_cast<View *>(&inventoryview),
+        dynamic_cast<View *>(&upgradeview),
         dynamic_cast<View *>(&achievementlistview),
         dynamic_cast<View *>(&optionsview),
     };
@@ -48,9 +51,9 @@ public:
     void render() override;
     const std::string &name() override { return _name; };
     void handleInput(const std::string &input) override;
-    explicit MainView(notifyMessage *msg, Wallet *wallet, Inventory *inventory, Store *store) :
+    explicit MainView(notifyMessage *msg, Wallet *wallet, Inventory *inventory, ItemStore *store) :
         wallet(wallet), msg(msg),
-        inventory(inventory), store(store) {
+        inventory(inventory), itemstore(store) {
         switchActiveView("1");
     };