+++ src/Screen/src/OptionsView.h
@@ -1,10 +1,10 @@
 #ifndef C_OOKIECLIKER_OPTIONSVIEW_H
 #define C_OOKIECLIKER_OPTIONSVIEW_H
-#include "View.h"
-#include "notifyMessage.h"
-#include "Store.h"
 #include "Inventory.h"
+#include "ItemStore.h"
+#include "View.h"
 #include "Wallet.h"
+#include "notifyMessage.h"
 
 class OptionsView : public View
 {
@@ -12,7 +12,7 @@ class OptionsView : public View
     notifyMessage *msg;
     Wallet *wallet;
     Inventory *inventory;
-    Store *store;
+    ItemStore *store;
     const std::string &savefilename;
     void save();
     void load();
@@ -21,7 +21,7 @@ public:
     void render() override;
     const std::string &name() override { return _name; };
     void handleInput(const std::string &input) override;
-    OptionsView(notifyMessage *msg, Wallet *wallet, Inventory *inventory, Store *store,
+    OptionsView(notifyMessage *msg, Wallet *wallet, Inventory *inventory, ItemStore *store,
                 const std::string &savefilename) :
         msg(msg), wallet(wallet), inventory(inventory), store(store), savefilename(savefilename) {};
 };